edge/requirements.txt
paho-mqtt==1.6.1
pydantic==2.10.6
requests==2.32.3
